requirements.txt
requests
python-dotenv
rookiepy
browser-cookie3
google-api-python-client
google-auth
google-auth-oauthlib
playwright
